refactor: extract React hooks for data fetching
Extract data fetching logic from page components into reusable hooks:

- useTraces: ListTraces RPC + filter state + debounced search
- useTrace: GetTrace RPC + span tree building + span selection
- useDashboard: GetUsageSummary RPC + loading/error state

Supporting changes:
- New src/lib/api.ts: singleton ConnectRPC clients
- New src/types/traces.ts: shared TraceSummaryRow, TraceFilters
- Updated pages to use hooks (no change in behavior or UI)
