+++ ui/src/app/traces/[id]/page.tsx
@@ -1,130 +1,9 @@
 "use client";
 
-import { useEffect, useState } from "react";
 import { useParams, useRouter } from "next/navigation";
-import { createClient } from "@connectrpc/connect";
-import { transport } from "@/lib/connect";
-import { TraceService } from "@/gen/v1/trace_service_pb";
-import type { Span } from "@/gen/types/trace_pb";
-import { SpanKind, SpanStatus } from "@/gen/types/trace_pb";
-
-// ──────────────────────────────────────────
-// Types
-// ──────────────────────────────────────────
-
-interface SpanNode {
-  span: Span;
-  children: SpanNode[];
-  depth: number;
-  /** Offset from trace start in ms */
-  offsetMs: number;
-  /** Duration in ms */
-  durationMs: number;
-}
-
-interface TraceData {
-  traceId: string;
-  rootSpanName: string;
-  totalDurationMs: number;
-  totalTokens: number;
-  totalCostUsd: number;
-  spanCount: number;
-  environment: string;
-  startTime: string;
-  tree: SpanNode[];
-  flatSpans: SpanNode[];
-}
-
-// ──────────────────────────────────────────
-// Helpers
-// ──────────────────────────────────────────
-
-function tsToMs(ts?: { seconds: bigint; nanos: number }): number {
-  if (!ts) return 0;
-  return Number(ts.seconds) * 1000 + (ts.nanos ?? 0) / 1e6;
-}
-
-function durationToMs(d?: { seconds: bigint; nanos: number }): number {
-  if (!d) return 0;
-  return Number(d.seconds) * 1000 + (d.nanos ?? 0) / 1e6;
-}
-
-function kindLabel(kind: SpanKind): string {
-  const map: Record<SpanKind, string> = {
-    [SpanKind.UNSPECIFIED]: "span",
-    [SpanKind.LLM]: "LLM",
-    [SpanKind.AGENT]: "Agent",
-    [SpanKind.TOOL]: "Tool",
-    [SpanKind.RETRIEVAL]: "RAG",
-    [SpanKind.EMBEDDING]: "Embed",
-    [SpanKind.CHAIN]: "Chain",
-    [SpanKind.GENERAL]: "General",
-  };
-  return map[kind] ?? "span";
-}
-
-function kindColor(kind: SpanKind): string {
-  const map: Record<SpanKind, string> = {
-    [SpanKind.UNSPECIFIED]: "var(--text-secondary)",
-    [SpanKind.LLM]: "#a78bfa",
-    [SpanKind.AGENT]: "#f59e0b",
-    [SpanKind.TOOL]: "#34d399",
-    [SpanKind.RETRIEVAL]: "#60a5fa",
-    [SpanKind.EMBEDDING]: "#f472b6",
-    [SpanKind.CHAIN]: "#fb923c",
-    [SpanKind.GENERAL]: "var(--text-secondary)",
-  };
-  return map[kind] ?? "var(--text-secondary)";
-}
-
-function buildTree(spans: Span[], traceStartMs: number): SpanNode[] {
-  const nodeMap = new Map<string, SpanNode>();
-
-  // Create nodes
-  for (const span of spans) {
-    nodeMap.set(span.spanId, {
-      span,
-      children: [],
-      depth: 0,
-      offsetMs: tsToMs(span.startTime) - traceStartMs,
-      durationMs: durationToMs(span.duration),
-    });
-  }
-
-  // Build tree
-  const roots: SpanNode[] = [];
-  for (const node of nodeMap.values()) {
-    if (node.span.parentSpanId && nodeMap.has(node.span.parentSpanId)) {
-      const parent = nodeMap.get(node.span.parentSpanId)!;
-      parent.children.push(node);
-    } else {
-      roots.push(node);
-    }
-  }
-
-  // Set depths and sort children by start time
-  function setDepth(node: SpanNode, depth: number) {
-    node.depth = depth;
-    node.children.sort((a, b) => a.offsetMs - b.offsetMs);
-    for (const child of node.children) {
-      setDepth(child, depth + 1);
-    }
-  }
-  roots.sort((a, b) => a.offsetMs - b.offsetMs);
-  for (const root of roots) setDepth(root, 0);
-
-  return roots;
-}
-
-function flattenTree(nodes: SpanNode[]): SpanNode[] {
-  const result: SpanNode[] = [];
-  function walk(node: SpanNode) {
-    result.push(node);
-    for (const child of node.children) walk(child);
-  }
-  for (const n of nodes) walk(n);
-  return result;
-}
+import { useTrace, kindLabel, kindColor } from "@/hooks/useTrace";
+import type { SpanNode } from "@/hooks/useTrace";
+import { SpanStatus } from "@/gen/types/trace_pb";
 
 // ──────────────────────────────────────────
 // Components
@@ -153,13 +32,9 @@ function SpanRow({
       role="button"
       tabIndex={0}
     >
-      {/* Left: span label with tree indentation */}
       <div className="waterfall-label" style={{ paddingLeft: node.depth * 20 + 8 }}>
         <span className="waterfall-connector" />
-        <span
-          className="waterfall-kind"
-          style={{ color, borderColor: color }}
-        >
+        <span className="waterfall-kind" style={{ color, borderColor: color }}>
           {kindLabel(node.span.kind)}
         </span>
         <span className="waterfall-name" title={node.span.name}>
@@ -168,7 +43,6 @@ function SpanRow({
         {isError && <span className="badge badge-error" style={{ marginLeft: 6, fontSize: 10 }}>ERR</span>}
       </div>
 
-      {/* Right: waterfall bar */}
       <div className="waterfall-bar-container">
         <div
           className="waterfall-bar"
@@ -284,7 +158,6 @@ function SpanDetail({ node }: { node: SpanNode }) {
             )}
           </div>
 
-          {/* Prompt / Completion */}
           {genAi.inputContent && (
             <div className="span-content-block">
               <div className="span-content-label">📥 Prompt</div>
@@ -358,47 +231,8 @@ export default function TraceDetailPage() {
   const router = useRouter();
   const traceId = params.id as string;
 
-  const [trace, setTrace] = useState<TraceData | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [selectedSpanId, setSelectedSpanId] = useState<string | null>(null);
-
-  useEffect(() => {
-    const client = createClient(TraceService, transport);
-    client
-      .getTrace({ traceId })
-      .then((res) => {
-        if (!res.trace) {
-          setError("Trace not found");
-          return;
-        }
-        const t = res.trace;
-        const traceStartMs = tsToMs(t.startTime);
-        const tree = buildTree(t.spans, traceStartMs);
-        const flatSpans = flattenTree(tree);
-
-        setTrace({
-          traceId: t.traceId,
-          rootSpanName: t.rootSpanName || "unknown",
-          totalDurationMs: durationToMs(t.duration),
-          totalTokens: Number(t.totalTokens) || 0,
-          totalCostUsd: t.totalCostUsd || 0,
-          spanCount: t.spanCount || t.spans.length,
-          environment: t.environment || "—",
-          startTime: t.startTime
-            ? new Date(traceStartMs).toLocaleString()
-            : "—",
-          tree,
-          flatSpans,
-        });
-      })
-      .catch((err) => setError(err.message))
-      .finally(() => setLoading(false));
-  }, [traceId]);
-
-  const selectedNode = trace?.flatSpans.find(
-    (n) => n.span.spanId === selectedSpanId
-  );
+  const { trace, loading, error, selectedSpanId, selectedNode, toggleSpan } =
+    useTrace(traceId);
 
   return (
     <>
@@ -481,7 +315,6 @@ export default function TraceDetailPage() {
 
             {/* Waterfall + Detail split */}
             <div className="waterfall-split">
-              {/* Waterfall panel */}
               <div className="waterfall-panel animate-in">
                 <div className="waterfall-header">
                   <span className="table-title">Span Waterfall</span>
@@ -513,13 +346,7 @@ export default function TraceDetailPage() {
                       node={node}
                       totalDurationMs={trace.totalDurationMs}
                       selected={node.span.spanId === selectedSpanId}
-                      onClick={() =>
-                        setSelectedSpanId(
-                          node.span.spanId === selectedSpanId
-                            ? null
-                            : node.span.spanId
-                        )
-                      }
+                      onClick={() => toggleSpan(node.span.spanId)}
                     />
                   ))}
                 </div>
